Script to generate per IPBES region summary
Uses logic similar to what I used in countrify.

Compute what the area-weighted mean intact BII (equivalent to the mean
cell area) and the ratio of the actual mean BII.

The vector file in OUTDIR/vector/ipbes_land_shape is a shapefile with
the definition of the the IPBES subregions.

vector is now the top-level folder for vector datasets.  Within it are
(for now) ipbes_land_shapes and ssp-regions (which I used in my MS
project).

Needs support for mismatched heights.  The projections for SSP5 have
slightly different bounds (perhaps because of population dynamics?)  and
this causes the bounds to not be aligned to the grid.  This somehow
causes the window for different rasters to be off by one.