Convert to use parallel xargs
Rather then trying to implement parallelism via & and wait, use the
parallel option of xargs.  This requires the loop be silent as all the
output gets sent to xargs.

I poked around a bit to try to figure out how to use a third file
descriptor but didn't figure it out right away and decided this was good
enough.