Jupyter notebook for visualizing indicator summary
Generates on chart for each BII indicator (BIIAb, BIISR) for each
scenario.  The list of scenarios is hardcoded (as a shortcut for now).

The data is loaded from the s3 bucket in the hope that it is then
possible to run the notebook from github (will depend on whether bokeh,
requests and other python packages are available).