Fix lat and lon array values
The file variables need to be the cell centers, not the upper-right
corner (as I originally had them).  This caused a weird problem were
gdal though the dataset was a "bottom-up" and would refuse to read it.

Change how to do the includes since the file now lives in scripts and
therefore is part of the module.