Add support for land use intensity in contrast coefficients
Moved as_contrast() from lu.luh5 to lui.luh5 since it is permissive of
UI terms in coefficients.

as_contrast() removed "Minimal" from the coefficient name.  I don't
remember why that was but it is no longer applicable since the RHS can
contain Minimal, e.g.

    ContrastPrimary Vegetation Minimal use-Secondary Vegetation Minimal use

is a valid name.  Since everything up-to the dash gets removed, it seems
harmless to keep Minimal.  But I need to run more tests to verify this.