Script to remove cells with overlap
When using all-touched there can be cells with overlap, i.e. they
appear in both the islands mask and the mainland mask.

This script removes (masks) those cells from a raster.  This allows
using all-touch when generating the mask but removing the cells that
overlap from the resulting raster.

When I tried generating the islands mask without all-touch, it resulted
in a raster with about 1M fewer cells.  That seemed like a lot.  So keep
all-touched but remove the cells with overlap.