Include the raster's mask when using it as a mask
RasterSet allows specifying a raster as a mask.  The idea being that you
can use one raster to decide which cells to evaluate.  This is not used
often but Katia had a need to do so recently.  I discovered that I was
reading the raster as masked and then checking for values.  Better to
read the raster masked, check for values and also ignore any cells that
are masked in the raster itself.