Use py-geospatial as the starting image for the container
Conda creates more problems than it solves.  So let's get rid of it.
This change moves to using my hand-rolled py-geospatial image as the
starting point for the container.  I then add all the required python
packages (all the underlying libraries should already be installed) and
copy over the requires sources and scripts.