Add support for generating a projected raster
When masking cells that have high BII in the present and the future,
things look a bit odd when unprojected because northern latitudes
dominate the image.

This change adds support for projecting the resulting raster using
cartopy.  Currently the projection is hard-coded as Mollweide.  For now
I also hard-coded to project the entire world, even regions that are not
in the results because otherwise the image looks weird.

I plan to add support for a command-line option to select cropping, but
the old version of rasterio (0.36) seems a bit broken when it comes to
boundless reads.