Container for running Xvnc in as a trampoline for X11 display
One problem with containers is how to get X11 working.  If ssh is not
running, there's no easy way to redirect X11 from within the container
to my host.  And since the R containers are not based on an image that
has sshd running, this is a bit of a problem.

As a work-around I created this container which just runs a wide open
Xvnc server.  This allows me to remotely display stuff from within
containers and not make my display itself vulnerable.

It's a bit unwieldy but it works.