Limit scaling / data input to extent of projection
When using a small-area projection (like OSGB) it doesn't make sense to
downsample the data, since we want as many pixels as possible for the
output.

This change modifies rview so it will first determine the extent (in
degrees) of the projection and then decide how much to downsample (if at
all).  This allows generating high resolution maps for small areas, but
still have a reasonable sized global map.