Fix support for bounds in projection
Long ago I added support for specifying a bounding box (extent) for the
projection (for example to limit the extent when using the high
resolution input data).  But I didn't hook it up correctly.  This change
fixes things so it is possible to specify the bounding box.

This is useful when debugging something so I can run small projections
quickly.