Script for Dasgupta report projections
This is the first script I write since removing predictification.  Plus
it is significantly different from normal projections that I decided to
use a specialized script for it.

Instead of adding it at the top level, I'll add a project folder under
user-scripts with all the scripts for the project.