Add support for projecting based on WorldPop human population data
WorldPop is supposed to be a better (smoother) population estimate
dataset with less artifacts at administrative boundaries.  NatGeo seemed
interested in seeing the results of projections with it.  The results
where not as expected because the HPD data still has artifacts.

This commit adds support for using WorldPop data for NatGeo
projections.