Update to R 4.0.0
This commit upgrades the Dockerfiles to built on top of
rocker/tidyverse:4.0.0.  This required a small bit of fiddling with
packages (Ubuntu, python, and R) because it upgraded to Ubuntu focal and
a few packages have changed.

I haven't coped the rocker container build changes as it doesn't seem
worthwhile at the moment.