Finish lint cleanup
This second comment disables a number of warnings (for specific lines)
and cleans up the lu, lui, and hpd sub-modules.

Disabled warnings are unused imports (in __init__.py files), places
where an import after code (after matplotlib.use() call), complex
functions (that I don't want to break-up as part of this effort), and a
few places where I explicitly compare against False or None in a Numpy
function.

Here is a summary of the errors/warnings I allowed:

     15 C901
      4 E203
     13 E402
      6 E501
      7 E712
      2 E731
      7 F401
      4 F811
      1 F841