Re-implement change 9cf8f4a after refactoring
In all the refactoring of code around, the version history for
predicts.py got lost.  That meant I could not rebase the proj/manuscript
branch on top of docker (the file appears as deleted).

So I rebased the branch skipping change 9cf8f4a and then re-implemented
the change on top.