Add urbanize command to rasterize.py
This commit adds a new urbanize command that generates a raster for the
fraction of each cell that is "urban infrastructue" based on the
MapBiomas dataset.

It also fixes a problem with the CRS.  The land-use data data is in 0.5°
resolution so use the RCP rasters as a reference.  Also change where the
output files are stored.