Use the newly generated soil properties files
I wrote a script (soil-grids.py) to process the Soil Grid data.  The
script average the three depths, clips (by turning to NA) for (min, max)
range, and scales the units of the data (there are some differences in
the units of the Soil Grid data compared to Victoria's models).

Updated the projection script to use new soil properties.  Also updated
README file.